Split test files for lane
